Feat/#58 기능 추가 및 테스트 코드 작성

+++ RevUp-api/src/main/java/com/revup/answer/dto/response/AnswerDetailsResponse.java
@@ -21,7 +21,7 @@ public record AnswerDetailsResponse(
                 answer.getUser().getNickname(),
                 answer.getUser().getProfileImage(),
                 answer.getUser().getTotalAnswerCount(),
-                answer.getUser().getAdoptedRate(),
+                answer.getUser().getAdoptedRate().intValue(),
                 answer.getCreatedAt().toString(),
                 answer.getIsAccept().toBoolean()
         );
